Rename to optimuspy. finish best mode.
