+++ config.ini
@@ -5,3 +5,11 @@ user=Admin
 password=YXBwbGU=
 decode_b64=True
 ssl=True
+
+[tm1srv02]
+address=localhost
+port=12354
+user=Admin
+password=YXBwbGU=
+decode_b64=True
+ssl=True
